+++ packages/nano_domain/lib/src/nano_domain.dart
@@ -38,6 +38,7 @@ export 'social/friends_leaderboard.dart';
 export 'social/social_share.dart';
 export 'social/safety_report.dart';
 export 'social/moderation_queue.dart';
+export 'social/safety_policy.dart';
 export 'media/asset_review.dart';
 export 'media/generated_asset.dart';
 export 'media/generation_budget.dart';